+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/ModInitializerImpl.java
@@ -184,6 +184,7 @@ public class ModInitializerImpl extends AbstractModInitializer implements ModIni
         registerAdditionalFilteredResourceFactories();
         registerDestructorStrategyFactories();
         registerConstructorStrategyFactories();
+        registerAlternativeGridHints();
         registerAdditionalGridInsertionStrategyFactories();
         registerGridExtractionStrategyFactories();
         registerGridScrollingStrategyFactories();