+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/ModInitializer.java
@@ -197,6 +197,7 @@ public class ModInitializer extends AbstractModInitializer {
         registerAdditionalFilteredResourceFactories();
         registerDestructorStrategyFactories();
         registerConstructorStrategyFactories();
+        registerAlternativeGridHints();
         registerAdditionalGridInsertionStrategyFactories();
         registerGridExtractionStrategyFactories();
         registerGridScrollingStrategyFactories();